wip: :construction:  move modules of flow as etude
